+++ tests/on-device/config-android-integration.nix
@@ -6,6 +6,7 @@ _:
     am.enable = true;
     termux-open.enable = true;
     termux-open-url.enable = true;
+    termux-reload-settings.enable = true;
     termux-setup-storage.enable = true;
     termux-wake-lock.enable = true;
     termux-wake-unlock.enable = true;