+++ tests/on-device/config-android-integration.nix
@@ -1,0 +1,8 @@
+_:
+
+{
+  system.stateVersion = "23.11";
+  android-integration = {
+    am.enable = true;
+  };
+}