+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * New options under `android-integration`,
   offer some of the tools familiar to Termux users.
-  Currently it's just `am`.
+  Currently only `am` and `termux-setup-storage` are provided.
 
 ### Compatibility considerations