+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 * New options under `android-integration`,
   offer some of the tools familiar to Termux users:
-  `am`, `termux-open`, `termux-open-url`, `termux-setup-storage` and `xdg-open`.
+  `am`, `termux-open`, `termux-open-url`,
+  `termux-setup-storage`, `termux-wake-lock`, `termux-wake-unlock`
+  and `xdg-open`.
 
 ### Compatibility considerations