+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Release 24.05 (unreleased)
+## Release 24.11 (unreleased)
+
+## Release 24.05
 
 ### New Options